requirements.txt
numpy>=1.24.3
opencv-python>=4.8.0
pyttsx3>=2.90
SpeechRecognition>=3.10.0
openai>=0.27.8
pyaudio>=0.2.13
matplotlib>=3.7.1
pyyaml>=6.0.1
python-socketio==5.11.0
pydub==0.25.1
pyserial==3.5
pygame>=2.5.0
